SWE63: Fixed date picker, priviledge bugs, and removed unneeded db fields

+++ src/main/java/com/unb/tracker/web/CourseController.java
@@ -125,6 +125,10 @@ public class CourseController {
     @PostMapping(value = "courses/{courseId}/delete")
     @ResponseBody
     public String deleteCourse(@PathVariable Long courseId, Principal principal) {
+        User user = userRepository.findByUsername(principal.getName());
+        if (user == null || !user.getHasExtendedPrivileges()) {
+            throw new BadRequestException();
+        }
         courseRepository.delete(courseId);
         return "success";
     }
